Fix bug - mada payment
Show egypt payout methods in egypt only

Allow fractions for transaction amount

+++ database/migrations/2020_03_24_045604_create_transactions_table.php
@@ -19,7 +19,7 @@ class CreateTransactionsTable extends Migration
             $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->tinyInteger('type')->index();
-            $table->integer('amount');
+            $table->decimal('amount', 20, 2);
             $table->decimal('amount_usd', 20, 4)->nullable();
             $table->tinyInteger('payment_method')->index();
             $table->tinyInteger('status')->index();
